Decoupled CommunicatorTester from SlaveTester
